% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_match.R
\name{area_match}
\alias{area_match}
\title{Function that approximately matches shape file areas through area. From Arran's area match function}
\usage{
area_match(shp1_old, shp1_new)
}
\arguments{
\item{shp1_old}{Shape files of old gadm system}

\item{shp1_new}{Shape files of new gadm system}

\item{increment}{Width of the grid squares. Defaults to 0.15}
}
\value{
proportion of old shapefiles in new shapefiles
}
\description{
Function that approximately matches shape file areas through area. From Arran's area match function
}
