% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_match.R
\name{grid_match_multi}
\alias{grid_match_multi}
\title{grid match for multiple countries}
\usage{
grid_match_multi(shp1_old, shp1_new, increment = 0.15)
}
\arguments{
\item{shp1_old}{Shape files of old gadm system}

\item{shp1_new}{Shape files of new gadm system}

\item{increment}{Width of the grid squares. Defaults to 0.15}
}
\value{
grid match for multiple countries and list of unmatched countries
}
\description{
grid match for multiple countries
}
