% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shp_dataframes.R
\name{make_shp_dataframes}
\alias{make_shp_dataframes}
\title{Make dataframes of shape files}
\usage{
make_shp_dataframes(shp1_new, shp1_old)
}
\arguments{
\item{shp1_new}{Shape files of new gadm system}

\item{shp2_new}{Shape files of old gadm system}
}
\value{
dataframes of shape file info
}
\description{
Make dataframes of shape files
}
