% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gadm.R
\name{match_gadm}
\alias{match_gadm}
\title{Match gadm main function}
\usage{
match_gadm(shp1_new, shp1_old, match_nearest)
}
\arguments{
\item{shp1_new}{Shape files of new gadm system}

\item{shp1_old}{Shape files of old gadm system}

\item{match_nearest}{whether to match the remaining entries to nearest centroid}
}
\value{
dataframe of old to new adm translation
}
\description{
Match gadm main function
}
