% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gadm_by_address.R
\name{match_gadm_by_address}
\alias{match_gadm_by_address}
\title{Match gadm by their address}
\usage{
match_gadm_by_address(joined_df, from = "x", to = "y", dat)
}
\arguments{
\item{joined_df}{data frame of joined shape dataframes}

\item{from}{which way to match from}

\item{to}{which way to match to}
}
\value{
updated dataframe of match and list of unmatched names
}
\description{
Match gadm by their address
}
