% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gadm_by_name_poly.R
\name{match_gadm_by_name_and_poly}
\alias{match_gadm_by_name_and_poly}
\title{Match gadm by name and polygon}
\usage{
match_gadm_by_name_and_poly(joined_df, from = "x", to = "y", dat, shp1)
}
\arguments{
\item{joined_df}{data frame of joined shape dataframes}

\item{from}{which way to match from}

\item{to}{which way to match to}
}
\value{
updated dataframe
}
\description{
Match gadm by name and polygon
}
