% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gadm_to_nearest_centroid.R
\name{match_gadm_to_nearest_centroid}
\alias{match_gadm_to_nearest_centroid}
\title{Match gadm to the nearest centroid}
\usage{
match_gadm_to_nearest_centroid(joined_df, from = "x", to = "y", dat)
}
\arguments{
\item{joined_df}{data frame of joined shape dataframes}

\item{from}{which way to match from}

\item{to}{which way to match to}

\item{dat}{dataframe to be matched to}
}
\value{
updated dataframe
}
\description{
Match gadm to the nearest centroid
}
