% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGADM-package.R
\docType{package}
\name{matchGADM-package}
\alias{matchGADM}
\alias{matchGADM-package}
\title{matchGADM: Converts between shape files}
\description{
This package contains two main functions which both map from 
    one shapefile to another through matching the area. The resulting output 
    is a data frame of proportions from each admin unit in the old to the new.
    This can be used to convert population sizes from one map to another.
}
\author{
\strong{Maintainer}: Arran Hamlet \email{arran.hamlet14@imperial.ac.uk}

Authors:
\itemize{
  \item Katy Gaythorpe
}

}
\keyword{internal}
