% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGADM.R
\docType{package}
\name{matchGADM}
\alias{matchGADM}
\alias{matchGADM-package}
\title{matchGADM: A package for matching old and new versions of gadm.}
\description{
matchGADM: A package for matching old and new versions of gadm.
}
