% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_by_lonlat.R
\name{merge_by_lonlat}
\alias{merge_by_lonlat}
\title{Merge dataframes of shapefiles by rounded longitude and latitude}
\usage{
merge_by_lonlat(dat_old, dat_new)
}
\arguments{
\item{dat_old}{dataframe of old shapefiles}

\item{dat_new}{dataframe of new shapefiles}
}
\value{
data frame of joined shape dataframes
}
\description{
Merge dataframes of shapefiles by rounded longitude and latitude
}
