% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_shp.R
\name{tidy_shp}
\alias{tidy_shp}
\title{Tidy shapefiles}
\usage{
tidy_shp(shp1_new, shp1_old)
}
\arguments{
\item{shp1_new}{Shape files of new gadm system}

\item{shp1_old}{Shape files of old gadm system}
}
\value{
Cleaned versions of the shapefiles
}
\description{
Tidy shapefiles
}
