% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_match.R
\name{grid_match}
\alias{grid_match}
\title{Function that approximately matches shape file areas through a grid of points}
\usage{
grid_match(shp1_old, shp1_new, increment = 0.15)
}
\arguments{
\item{shp1_old}{Shape files of old gadm system}

\item{shp1_new}{Shape files of new gadm system}

\item{increment}{Width of the grid squares. Defaults to 0.15}
}
\value{
proportion of old shapefiles in new shapefiles (approximate)
}
\description{
Function that approximately matches shape file areas through a grid of points
}
