% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGADM-package.R
\docType{package}
\name{matchGADM-package}
\alias{matchGADM}
\alias{matchGADM-package}
\title{matchGADM: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
